requirements.txt
pandas
ollama
python-dotenv
requests